% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERGM_result_plot_function.R
\name{ERGM_result_plot}
\alias{ERGM_result_plot}
\title{ERGM_result_plot}
\usage{
ERGM_result_plot(ergm.fit, custom_var_names, plot_title, y_axis_lab = TRUE)
}
\arguments{
\item{ergm.fit}{ERGM object}

\item{custom_var_names}{Variable names}

\item{plot_title}{Plot title}

\item{y_axis_lab}{TRUE/FALSE - do you want the y-axis labels present. FALSE useful for panel plots. Default TRUE}
}
\value{
plot
}
\description{
creates plot for ERGM results
}
