% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlreg_convert_function.R
\name{htmlreg_convert}
\alias{htmlreg_convert}
\title{htmlreg_convert}
\usage{
htmlreg_convert(doc, excel_format)
}
\arguments{
\item{doc}{The .docx file}

\item{excel_format}{TRUE for excel file, FALSE for a csv}
}
\value{
writes the output as a table in excel or csv format.
}
\description{
This function converts a .docx file (created using texreg's htmlreg function) to a tabular file (excel or csv)
}
